Login Page Design
Added in the Login.aspx page. It has two text boxes, one for username
and one for password. There is validation to ensure that users enter in
values for the username/password.

Changed the menu links in the MasterPage to Hyperlinks for easier
alterations.

Added in the code necessary for validation in the Web.config file.

NOTE: Login redirects to HomePage but session control/username-password
still needs to be implemeneted.