Cinderella Check-In GridView Modified
The CinderellaListBox is now a GridView to allow for multiple columns.
It is databound to the Cinderella entity and has the Fname, Lname,
AppointmentDateTime, and ID (not visible) of the Cinderellas.

It is sortable by selecting the headers. To select a row a user must
click on the 'select' links that are located on the left of every row.
The view has a set height and is scrollable.

Error-checking fluff is entered for the Checkin button.