AppointmentManagement Functionality
ManageAppointment page now has SQL functionality. The selected
Cinderella's AppointmentTime will be UPDATED with the selected Date
(today's date set as default) with the selected Time (from the three
drop down boxes).

There is validation to ensure the user selects a Cinderella, and a
notification message to show that the new appointment time has been
set.