Checkout Functionality/Status Adjustments
Add in functionality to the Checkout Page. It takes the package
information the user enters and makes a Package with the listed
information and the selected Cinderella's ID. Sets the Package's
'Inpackaging' status to Y.

Sets the Cinderella's status to 'Waiting for Package' and the associated
Godmother's status to 'On Break'.

Also adjusted Checkin/ManualPairing code to now implement the 'Waiting
for ...'  terminology.

Deleted two images that are not being used.