ManagePackaging Layout
Added in the ManagePackaging Page to organize the Packaging station. The
Create Package page will create the Cinderella's package in the database
and give it the status of InPackaging = 'Y'.

Another Page, PackageCheckout (STILL TO BE CREATED), will allow users to
mark packages as Checked Out and ready to be delivered.

Changed the appropriate HyperLinks in the MasterPage.