VolunteerClocking Page and Functionality
Added in the Volunteer Clocking Page with Functionality. This page
features two tables, one with Volunteers On Break, and one with
Volunteers not On Break. By selecting a volunteer and clicking the
corresponding button under the table, the Volunteer will be taken off or
put on break.

Included HyperLink capability in the menu.

Also organized some of the SQL code for the MaualPairingPage.