Initial Volunteer Role and Shift Management
To VolunteerRegistration, added in SQL to INSERT ShiftRecords according
to the selected information.

To VolunteerCheckin, altered the DataGrid to only show Volunteers whose
ShiftName lines up with the current day. Additionally, when checking in
their RoleRecord is officially started with the Role they signed up for
in Registration. Volunteers that show up for days they did not register
for will have to go to the Volunteer Central to be entered in a new
Shift.

PLEASE NOTE: For testing purposes, the SQL code is only searching for
Volunteers whose Shifts line up with FRIDAY, one of the testing days
entered into the database. For actual implementation that checks with
the current day, use the code datename(dw,getdate()) in place of
'Friday' to check if the ShiftName lines up with the current day of the
week.