Login/log out capability, form tag removal
- Edit log in page so that any user can log on.
- Manage Session based on account that is used
- Edit master page so that master page display reflects on user account
type.
- Created functions to handle master page based on account type.
- Inserted MasterPageLayout function into each PageLoad event in
solution
- Added log out out capabilities.
- Added welcome label.
- Removed form tags from every page. This is due to the LabelButton
added on the master page. Because of the LabelButton a form tag was
created on the Master page which resulted in a conflict (form tags
cannot be nested). Since every page is a slave to the master, there is
no need for a form tag in each page since each slave page references the
form tag on the master page.