Alterstions and CreatePackage modifications
Modified Alterations and CreatePackage to work with the correct
DressSize and DressLength (database changed as well to accommodate those
(values), along with ShoeSize and ShoeColor for CreatePackage.

Additionally reorganized items on the two pages to better manage
information via nesting tables which apparently is possible.

Factored in the ReadyForPickup value for Alterations.

Made Alterations only enable the necessary items depending on the
selected grid.