Explanation of "Major Alterations Changes!"
Accidentally commited changes before providing a summary, this is an
explanation of the previous commit:

Alterations is now one page once again (no Alterations Manager Pager)

Alterations now handles both the INSERTion and UPDATEion of all
alterations entities. This INCLUDES marking an alteration's dress as
ready to be picked up.

Functionality includes: differentiating dresses slected from the
GridView as already having an Alteration (and needing an UPDATE) or not
having an Alteration (and needing an INSERT)

Additionally, included functionality for the Seamstress DropDown
DataSource (still need to error-check to verify a seamstress is
selected)

Still need a space in Package Checkout/Creation to notify users that
Dresses are ready to be picked up