EditPackage Functionality
Added Functionality to the Edit Package Page. This page provides a list
of all COMPLETED packages. Selecting a Cinderella shows their package
information and allows the user to enter in new information for Dresses,
Shoes, Jewelry, and Notes.

Note that a completed package is one where the dress is present. The
edit package page also provides a list of dresses that are ready to be
delivered from alterations. When a dress is successfully delivered, the
user will mark it as such by clicking the Dress Delivered button.