ManageShopping Page limited functionality
The ManageShoppping page now has a GridView of paired Cinderellas and
Volunteers. When selected, the Go Shopping button becomes available
which changes the selected targets' statuses to 'Shopping', thereby
moving them to a Currently Shopping GridView.

STILL NEED: Functionality of the Undo Shopping button
Everything to do with ManualPairing and breaking pairings