Modified ManageShopping Pair button to handle already paired Cinderellas/Volunteers
The Pair button now has logic to determine if the selected Cinderella
and/or Volunteer are already paired. The not-needed pairings are placed
back in the queue and have their information altered to be ready for
another pairing cycle.

Additionally changed VolunteerCheckin to enqueue the Volunteers to the
front of the queue